+++ src/intellisense.ts
@@ -2,12 +2,34 @@ import * as vscode from 'vscode';
 import * as cp from 'child_process';
 import * as path from 'path';
 import * as fs from 'fs';
-import { ensureDirectoryExists, logger, LogLevel } from './util';
+import { ensureDirectoryExists, logger, LogLevel, notUndefined } from './util';
 import { langId, binName } from './consts';
 
+
 // INTELLISENSE
+
+/// make ready an output channel
+let intellisenseOutputChannel: vscode.OutputChannel | undefined = undefined;
+const initIntellisense = () => {
+    intellisenseOutputChannel = vscode.window.createOutputChannel(`VSLilyPond: Intellisense`);
+};
+
+const outputToChannel = async (msg: string, show: boolean = false) => {
+    if (intellisenseOutputChannel) {
+        intellisenseOutputChannel.appendLine(msg);
+        if (show) {
+            intellisenseOutputChannel.show(true);
+        }
+    }
+    else {
+        logger(`Unable to output to Intellisense OutputChannel, ${msg}`, LogLevel.warning, true);
+    }
+};
+
+let intellisenseProcess: cp.ChildProcessWithoutNullStreams | undefined = undefined;
 let timeout: any = undefined;
-const triggerIntellisense = (doc: vscode.TextDocument, diagCol: vscode.DiagnosticCollection, context: vscode.ExtensionContext) => {
+
+const triggerIntellisense = async (doc: vscode.TextDocument, diagCol: vscode.DiagnosticCollection, context: vscode.ExtensionContext) => {
     if (timeout) {
         clearTimeout(timeout);
         timeout = undefined;
@@ -77,7 +99,7 @@ export const getDiagError = (errGroup: string[]): DiagErrorInfo => {
     };
 };
 
-const processIntellisenseErrors = (output: string, doc: vscode.TextDocument, diagCol: vscode.DiagnosticCollection) => {
+const processIntellisenseErrors = async (output: string, doc: vscode.TextDocument, diagCol: vscode.DiagnosticCollection) => {
     const errorGroups = groupErrors(output);
     const processErrorGroup = (errGroup: string[]): vscode.Diagnostic | undefined => {
         try {
@@ -98,60 +120,74 @@ const processIntellisenseErrors = (output: string, doc: vscode.TextDocument, dia
         }
     };
 
-    // @ts-ignore
     const gottenDiag: vscode.Diagnostic[] =
         errorGroups
-            .map(errGroup => processErrorGroup(errGroup))
-            .filter(x => x !== undefined);
+            .map(processErrorGroup)
+            .filter(notUndefined);
+
     const currentDiag = diagCol.get(doc.uri) ?? [];
 
     const diagnostics = currentDiag.concat(gottenDiag);
     diagCol.set(doc.uri, diagnostics);
 };
 
-const execIntellisense = (doc: vscode.TextDocument, diagCol: vscode.DiagnosticCollection, context: vscode.ExtensionContext) => {
-    diagCol.clear();
+const execIntellisense = async (doc: vscode.TextDocument, diagCol: vscode.DiagnosticCollection, context: vscode.ExtensionContext) => {
+    try {
+        diagCol.clear();
 
-    const tmpPath = context.storagePath ?? "lilypondTmp";
-
-    /// ensure this dir exists
-    ensureDirectoryExists(tmpPath);
-
-    const tmpFilePath = path.join(tmpPath, 'intellisenseTmp.txt');
-
-    const config = vscode.workspace.getConfiguration(`vslilypond`);
-
-    fs.writeFile(tmpFilePath, doc.getText(), (err) => {
-        if (err) {
-            /// mute here because it is just intellisense
-            logger(err.message, LogLevel.error, true);
-        }
-        else {
-            const additionalArgs: string[] = config.intellisense.additionalCommandLineArguments.trim().split(/\s+/);
-
-            /// to include the current directory of the file as a search path
-            const includeArg = `--include=${path.dirname(doc.uri.fsPath)}`;
+        const tmpPath = context.storagePath ?? "lilypondTmp";
     
-            const args = [`-s`].concat(additionalArgs).concat(includeArg).concat(tmpFilePath);
+        /// ensure this dir exists
+        ensureDirectoryExists(tmpPath);
     
-            const s = cp.spawn(binName, args, { cwd: tmpPath });
+        const tmpFilePath = path.join(tmpPath, 'intellisenseTmp.txt');
     
-            s.stdout.on('data', (data) => {
-                logger(`Intellisense: no errors, ${data}`, LogLevel.info, true);
-            });
+        const config = vscode.workspace.getConfiguration(`vslilypond`);
     
-            s.stderr.on('data', (data) => {
-                processIntellisenseErrors(data.toString(), doc, diagCol);
-            });
+        fs.writeFile(tmpFilePath, doc.getText(), (err) => {
+            if (err) {
+                /// mute here because it is just intellisense
+                logger(err.message, LogLevel.error, true);
+            }
+            else {
+                const additionalArgs: string[] = config.intellisense.additionalCommandLineArguments.trim().split(/\s+/);
     
-            s.on('close', (code) => {
-                logger(`Intellisense process exited with code ${code}`, LogLevel.info, true);
-            });
-        }
-    });
+                /// to include the current directory of the file as a search path
+                const includeArg = `--include=${path.dirname(doc.uri.fsPath)}`;
+    
+                const args = [`-s`].concat(additionalArgs).concat(includeArg).concat(tmpFilePath);
+    
+                if (intellisenseProcess) {
+                    intellisenseProcess.kill();
+                    intellisenseProcess = undefined;
+                }
+    
+                intellisenseProcess = cp.spawn(binName, args, { cwd: tmpPath });
+    
+                intellisenseProcess.stdout.on('data', (data) => {
+                    logger(`Intellisense: no errors, ${data}`, LogLevel.info, true);
+                });
+    
+                intellisenseProcess.stderr.on('data', (data) => {
+                    processIntellisenseErrors(data.toString(), doc, diagCol);
+                });
+    
+                intellisenseProcess.on('close', (code) => {
+                    logger(`Intellisense process exited with code ${code}`, LogLevel.info, true);
+                    intellisenseProcess = undefined;
+                });
+            }
+        });
+    }
+    catch (err) {
+        const errMsg = `Intellisense failed with error ${err.message}`;
+        logger(errMsg, LogLevel.error, true);
+        outputToChannel(errMsg, true);
+    }
 };
 
 export const subscribeIntellisense = (context: vscode.ExtensionContext, diagCol: vscode.DiagnosticCollection) => {
+    initIntellisense();
     if (vscode.window.activeTextEditor && vscode.window.activeTextEditor.document.languageId === langId) {
         triggerIntellisense(vscode.window.activeTextEditor.document, diagCol, context);
     }