+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.1.3
+- Optimise process spawning by limiting one each for compilation and intellisense
+- Code optimisations
+- Using output channels for failing processes to notify user
+- Added `Play from...` option for MIDI playback, allows user to input timestamp for MIDI to start from
+- Only show status bar items for MIDI playback & input when options are valid
+
 ## 1.1.2
 - Fix MIDI Input accepting messages that are not MIDI music messages (e.g. Modulation)